Move FpsMonitor from utils to video module
This commit relocates the FpsMonitor class from the utils module to the video module.
