--- docs/utils/fps.md
@@ -1,3 +1,0 @@
-## list_files_with_extensions
-
-:::supervision.utils.fps.FpsMonitor
