+++ docs/utils/video.md
@@ -6,6 +6,10 @@
 
 :::supervision.utils.video.VideoSink
 
+## FPSMonitor
+
+:::supervision.utils.video.FPSMonitor
+
 ## get_video_frames_generator
 
 :::supervision.utils.video.get_video_frames_generator
@@ -13,3 +17,5 @@
 ## process_video
 
 :::supervision.utils.video.process_video
+
+
