+++ supervision/__init__.py
@@ -44,7 +44,6 @@ from supervision.geometry.utils import get_polygon_center
 from supervision.metrics.detection import ConfusionMatrix, MeanAveragePrecision
 from supervision.tracker.byte_tracker.core import ByteTrack
 from supervision.utils.file import list_files_with_extensions
-from supervision.utils.fps import FpsMonitor
 from supervision.utils.image import ImageSink, crop_image
 from supervision.utils.notebook import plot_image, plot_images_grid
 from supervision.utils.video import (
@@ -52,4 +51,5 @@ from supervision.utils.video import (
     VideoSink,
     get_video_frames_generator,
     process_video,
+    FPSMonitor
 )
